And justice for all

+++ BSOPM_new/BSOPM_new/AnSolution.h
@@ -1,6 +1,7 @@
 // Copyright 2018 Romanov Alexander
 
 #include "BSOPM.h"
+#include "omp.h"
 #ifndef ____BLACK_SHOLES_ANALYTICAL_SOLUTION____
 #define ____BLACK_SHOLES_ANALYTICAL_SOLUTION____
 
@@ -13,7 +14,7 @@ public:
 	void writeAllFairs			(int num_Threads, int N, float* pT, float* pK, float* pS0, float* pC, float r, float sig, std::vector<float>& T, std::vector<float>& K, std::vector<float>& S0, std::vector<float>& C);
 	void executeStockPrice		(int nsteps, int npaths, float time, float s0, float r, float sig, float* sbuffer, unsigned int seed, int indexGen);
 	float writeOneVersion		(int numVer, int num_Threads, int N, float* pT, float* pK, float* pS0, float* pC, float r, float sig, std::vector<float>& T, std::vector<float>& K, std::vector<float>& S0, std::vector<float>& C);
-	float simulateStockPriceAn	(int npaths, float s0, float r, float sig, float time, unsigned int seed, int indexGen, double workTime);
+	double simulateStockPriceAn	(int npaths, double s0, double r, double sig, double time, unsigned int seed, int indexGen, double workTime);
 	float simulateStockPriceAnOptedPar(long int npaths, float s0, float r, float sig, float time, unsigned int seed, int indexGen, bool is_optimized, int numthreds, double &workTime);
 
 	void baseVer		(float* pT, float* pK, float* pS0, float* pC, int nsamples, float r, float sig);
